"loginUser.php"
"loginUser.php" file has been pushed. 
this file contains one user defined function which is "logged_in()"
if user is already logged in, Don't show the loginUser.php page

the display_message() function is a user defined function to show messages
and validate_user_login() function is for checking the user if he's credentials matches with the database or not .

the rest code are just simple HTML codes, its a template and little modified.