helper_functions.php
in each file the config.php file has been included, so that the connection of the database can be maintained.

in helper_functions.php, many small small user defined functions are being generated so that while coding and making 
various types of other big functions, we will be able to use these small functions easily by calling these small functions.

 function last_id
this is the php built in function called "mysqli_insert_id"


function redirect($location)
this is the php built in function for redirecting page [header("Location: {$location} ")]


function query($sql)
This is the php built in function for database query connection which is [ mysqli_query($connection, $sql) ]


function confirm($result)
function for confirming database connection  [ mysqli_error($connection) ]


function escape_string($string)
This is the php built in function escapes special characters in a string for use in an SQL statement.
[ mysqli_real_escape_string($connection, $string) ] 



function clean($string)
This is the php built in function for cleaning The htmlentities() function converts characters to HTML entities.
htmlentities($string)




function fetch_array($result)
function for fetching
This is the php built in function
[ mysqli_fetch_array($result) ]



function row_count($result)
function The mysqli_num_rows() function returns the number of rows in a result set.
mysqli_num_rows($result)



function set_message($msg)
function for setting up custom messages in sessions
in this function we used session to set the message to show 



function display_message()
function for displaying messages in sessions
in this function we used session to display the message to show. 
if a message is being set using the set_message function. then this display_message function will
show the message in that html page