frontend_functions.php
This functions are used in index.php file

function get_products()
we have used Delimeter. 
delimeter is used here to write html codes in a php file. 
within this delimeter all the static html codes are written here.
this html codes will show in the index.php file. and this will show all the pictures in the index.php file. 
this function get_products() has been placed in index.php file. so that, while index.php is being entered in the URL, then this 
function get_products() will be called and all the html pictures will be shown on the page.


function send_message()
this is a function for emailing.
this is a 3rd party library function called phpmailer. 
all the necessary codes are written directly from phpmailer websites.
we have set our own mailer address, mailer password, 


when the function send_message() is being called then an Email will go to that specific email address.