functions.php
function get_active_slide()   and    function get_slides()
these are the two functions that will create the jQuery slides in index.php 
at first these functions are in slider.php
then slider.php is included in index.php file

the pictures are in the templates. 
the helper function display_image is used here to show the image.