login_functions_users.php
This file is fully used for the login,registration, logout purposes.
all the functions here are for those purposes.

//displaying validation_errors for forms using bootstrap alert component
function validation_errors($error_message)
this function is made for showing the error messages in the page



//email exists
function email_exists($email)
this function is for, if the user is trying to register with an existing email id, then this email_exists will be called and it will check whether the email is existed or not.



//username exists
function username_exists($username)
this function is for, if the user is trying to register with an existing username, then this username_exists will be called and it will check whether the username is existed or not.



//integer number checking for phone and zipCode
function int_check($result)
this function is for the phone and zipCode fields, only numeric value will be accepted. alphabetic value won't be accepted.
php built in function is_numeric($result) is used here



//function for keeping the value if user makes a mistake in validating form. used in register.php
function keep_value($result)
this function is for all the fields. if the user submits the form but there is a mistake in it, then when the page gets refreshed, all the values in the field will be present. this is used for user efficiency.
the "isset" function is used here 





//function for encrypting id
//placed in recover.php
//we want to make sure the data is being submitted in the the token given page alone
//whatever the data getting comes from the form
//this is not going to be available for us unless we submit the form
//creates encrypted, UniqueID, Random token, gives everytime a new value
function token_generator()
md5(uniqid(mt_rand() is used here 




//placed in register.php
// checking for all the registration fields are OK
function validate_user_registration()
this function will check all the fields if valid or not. 
also, its secure by htmlentities() function





function register_user()
this function will take all the values from the user given inputs from all the fields from registration. 
it will  escapes special characters by escape_string() function
it will check email and username if its already exists or not
it will generate md5 and microtime password in database
also it will send an email




//placed in activate.php
//function for activate the user by a link given in email after the registration
function activate_user()
after the registration, an email will go to the user's email address, then after clicking that link, this function will check if everything is OK or not. it will check from database by using query, rowCount,confirm functions (helper functions)




//placed in loginUser.php
// checking for all the login fields are OK
function validate_user_login()
when a user is trying to login, this function will check email and pass from database to validate. 





function login_user()
it will check user given password and database password 
it will decrypt the database password and it will match it



//logging out from the cookie used in logout.php
function logged_out()
used session against email. if logged out. it will unset the session and destroy the session



//placed in recover.php
//function for recovering users password, by Forget Password
function recover_password_email()
token will generate against a session
SQL query will check the email and validation code  from the database



//placed in code.php
//fuction for validating the code that is given to user in email, forget password
function validation_code()



//function for reseting the password
function password_reset()
new pass  and confirm pass fields checks 


//Sending Email function for register_user(), recover_password()
function send_email()
this function is for sending emails.
directly from PHPmailer